+++ android/src/main/java/com/twiliorn/library/CustomTwilioVideoViewManager.java
@@ -20,7 +20,13 @@ import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_DATATRACK_MES
 import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_DATA_CHANGED;
 import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_DISCONNECTED;
 import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_DOMINANT_SPEAKER_CHANGED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_LOCAL_AUDIO_TRACK_PUBLICATION_FAILED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_LOCAL_AUDIO_TRACK_PUBLISHED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_LOCAL_DATA_TRACK_PUBLICATION_FAILED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_LOCAL_DATA_TRACK_PUBLISHED;
 import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_LOCAL_PARTICIPANT_SUPPORTED_CODECS;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_LOCAL_VIDEO_TRACK_PUBLICATION_FAILED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_LOCAL_VIDEO_TRACK_PUBLISHED;
 import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_NETWORK_QUALITY_LEVELS_CHANGED;
 import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_PARTICIPANT_ADDED_AUDIO_TRACK;
 import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_PARTICIPANT_ADDED_DATA_TRACK;
@@ -36,6 +42,17 @@ import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_PARTICIPANT_R
 import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_PARTICIPANT_REMOVED_VIDEO_TRACK;
 import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_RECONNECTED;
 import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_RECONNECTING;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_RECORDING_STARTED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_RECORDING_STOPPED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_REMOTE_AUDIO_TRACK_PUBLISHED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_REMOTE_AUDIO_TRACK_SUBSCRIPTION_FAILED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_REMOTE_AUDIO_TRACK_UNPUBLISHED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_REMOTE_DATA_TRACK_PUBLISHED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_REMOTE_DATA_TRACK_SUBSCRIPTION_FAILED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_REMOTE_DATA_TRACK_UNPUBLISHED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_REMOTE_VIDEO_TRACK_PUBLISHED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_REMOTE_VIDEO_TRACK_SUBSCRIPTION_FAILED;
+import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_REMOTE_VIDEO_TRACK_UNPUBLISHED;
 import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_SCREEN_SHARE_CHANGED;
 import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_STATS_RECEIVED;
 import static com.twiliorn.library.CustomTwilioVideoView.Events.ON_VIDEO_CHANGED;
@@ -63,6 +80,7 @@ public class CustomTwilioVideoViewManager extends SimpleViewManager<CustomTwilio
     private static final int RELEASE_RESOURCE = 10;
     private static final int TOGGLE_BLUETOOTH_HEADSET = 11;
     private static final int SEND_STRING = 12;
+    private static final int SEND_BINARY = 18;
     private static final int PUBLISH_VIDEO = 13;
     private static final int PUBLISH_AUDIO = 14;
     private static final int SET_REMOTE_AUDIO_PLAYBACK = 15;
@@ -148,6 +166,9 @@ public class CustomTwilioVideoViewManager extends SimpleViewManager<CustomTwilio
             case SEND_STRING:
                 view.sendString(args.getString(0));
                 break;
+            case SEND_BINARY:
+                view.sendBinary(args.getString(0));
+                break;
             case PUBLISH_VIDEO:
                 view.publishLocalVideo(args.getBoolean(0));
                 break;
@@ -211,6 +232,32 @@ public class CustomTwilioVideoViewManager extends SimpleViewManager<CustomTwilio
                 ON_RECONNECTING, MapBuilder.of("registrationName", ON_RECONNECTING),
                 ON_RECONNECTED, MapBuilder.of("registrationName", ON_RECONNECTED),
                 ON_DATA_CHANGED, MapBuilder.of("registrationName", ON_DATA_CHANGED)));
+        map.putAll(MapBuilder.of(
+                ON_RECORDING_STARTED, MapBuilder.of("registrationName", ON_RECORDING_STARTED),
+                ON_RECORDING_STOPPED, MapBuilder.of("registrationName", ON_RECORDING_STOPPED)));
+
+        map.putAll(MapBuilder.of(
+                ON_LOCAL_AUDIO_TRACK_PUBLISHED, MapBuilder.of("registrationName", ON_LOCAL_AUDIO_TRACK_PUBLISHED),
+                ON_LOCAL_AUDIO_TRACK_PUBLICATION_FAILED, MapBuilder.of("registrationName", ON_LOCAL_AUDIO_TRACK_PUBLICATION_FAILED),
+                ON_LOCAL_VIDEO_TRACK_PUBLISHED, MapBuilder.of("registrationName", ON_LOCAL_VIDEO_TRACK_PUBLISHED),
+                ON_LOCAL_VIDEO_TRACK_PUBLICATION_FAILED, MapBuilder.of("registrationName", ON_LOCAL_VIDEO_TRACK_PUBLICATION_FAILED),
+                ON_LOCAL_DATA_TRACK_PUBLISHED, MapBuilder.of("registrationName", ON_LOCAL_DATA_TRACK_PUBLISHED),
+                ON_LOCAL_DATA_TRACK_PUBLICATION_FAILED, MapBuilder.of("registrationName", ON_LOCAL_DATA_TRACK_PUBLICATION_FAILED)));
+
+        map.putAll(MapBuilder.of(
+                ON_REMOTE_AUDIO_TRACK_PUBLISHED, MapBuilder.of("registrationName", ON_REMOTE_AUDIO_TRACK_PUBLISHED),
+                ON_REMOTE_AUDIO_TRACK_UNPUBLISHED, MapBuilder.of("registrationName", ON_REMOTE_AUDIO_TRACK_UNPUBLISHED),
+                ON_REMOTE_AUDIO_TRACK_SUBSCRIPTION_FAILED, MapBuilder.of("registrationName", ON_REMOTE_AUDIO_TRACK_SUBSCRIPTION_FAILED)));
+
+        map.putAll(MapBuilder.of(
+                ON_REMOTE_VIDEO_TRACK_PUBLISHED, MapBuilder.of("registrationName", ON_REMOTE_VIDEO_TRACK_PUBLISHED),
+                ON_REMOTE_VIDEO_TRACK_UNPUBLISHED, MapBuilder.of("registrationName", ON_REMOTE_VIDEO_TRACK_UNPUBLISHED),
+                ON_REMOTE_DATA_TRACK_PUBLISHED, MapBuilder.of("registrationName", ON_REMOTE_DATA_TRACK_PUBLISHED),
+                ON_REMOTE_DATA_TRACK_SUBSCRIPTION_FAILED, MapBuilder.of("registrationName", ON_REMOTE_DATA_TRACK_SUBSCRIPTION_FAILED),
+                ON_REMOTE_DATA_TRACK_UNPUBLISHED, MapBuilder.of("registrationName", ON_REMOTE_DATA_TRACK_UNPUBLISHED)));
+
+        map.putAll(MapBuilder.of(
+                ON_REMOTE_VIDEO_TRACK_SUBSCRIPTION_FAILED, MapBuilder.of("registrationName", ON_REMOTE_VIDEO_TRACK_SUBSCRIPTION_FAILED)));
 
         return map;
     }
@@ -228,6 +275,7 @@ public class CustomTwilioVideoViewManager extends SimpleViewManager<CustomTwilio
                 .put("toggleRemoteSound", TOGGLE_REMOTE_SOUND)
                 .put("toggleBluetoothHeadset", TOGGLE_BLUETOOTH_HEADSET)
                 .put("sendString", SEND_STRING)
+                .put("sendBinary", SEND_BINARY)
                 .put("toggleScreenSharing", TOGGLE_SCREEN_SHARING)
                 .put("toggleDataTrack", TOGGLE_DATA_TRACK)
                 .build();